+++ aws_quota/check/elb.py
@@ -7,8 +7,8 @@ from .quota_check import QuotaCheck, InstanceQuotaCheck, QuotaScope
 def get_albs(session: boto3.Session):
     return list(
         filter(
-            lambda lb: lb["Type"] == "application",
-            session.client("elbv2").describe_load_balancers()["LoadBalancers"],
+            lambda lb: lb['Type'] == 'application',
+            session.client('elbv2').describe_load_balancers()['LoadBalancers'],
         )
     )
 
@@ -16,8 +16,8 @@ def get_albs(session: boto3.Session):
 def get_nlbs(session: boto3.Session):
     return list(
         filter(
-            lambda lb: lb["Type"] == "network",
-            session.client("elbv2").describe_load_balancers()["LoadBalancers"],
+            lambda lb: lb['Type'] == 'network',
+            session.client('elbv2').describe_load_balancers()['LoadBalancers'],
         )
     )
 
@@ -26,39 +26,39 @@ class ClassicLoadBalancerCountCheck(QuotaCheck):
     key = "elb_clb_count"
     description = "Classic Load Balancers per region"
     scope = QuotaScope.REGION
-    service_code = "elasticloadbalancing"
-    quota_code = "L-E9E9831D"
+    service_code = 'elasticloadbalancing'
+    quota_code = 'L-E9E9831D'
 
     @property
     def current(self):
         return len(
-            self.boto_session.client("elb").describe_load_balancers()["LoadBalancerDescriptions"]
+            self.boto_session.client('elb').describe_load_balancers()['LoadBalancerDescriptions']
         )
 
 
 class ListenerPerClassicLoadBalancerCountCheck(InstanceQuotaCheck):
     key = "elb_listeners_per_clb"
     description = "Listeners per Classic Load Balancer"
-    service_code = "elasticloadbalancing"
-    quota_code = "L-1A491844"
-    instance_id = "Load Balancer Name"
+    service_code = 'elasticloadbalancing'
+    quota_code = 'L-1A491844'
+    instance_id = 'Load Balancer Name'
 
     @staticmethod
     def get_all_identifiers(session: boto3.Session) -> typing.List[str]:
         return [
-            lb["LoadBalancerName"]
-            for lb in session.client("elb").describe_load_balancers()["LoadBalancerDescriptions"]
+            lb['LoadBalancerName']
+            for lb in session.client('elb').describe_load_balancers()['LoadBalancerDescriptions']
         ]
 
     @property
     def current(self):
         try:
             return len(
-                self.boto_session.client("elb").describe_load_balancers(
+                self.boto_session.client('elb').describe_load_balancers(
                     LoadBalancerNames=[self.instance_id]
-                )["LoadBalancerDescriptions"][0]["ListenerDescriptions"]
+                )['LoadBalancerDescriptions'][0]['ListenerDescriptions']
             )
-        except self.boto_session.client("elb").exceptions.AccessPointNotFoundException as e:
+        except self.boto_session.client('elb').exceptions.AccessPointNotFoundException as e:
             raise InstanceWithIdentifierNotFound(self) from e
 
 
@@ -66,16 +66,16 @@ class NetworkLoadBalancerCountCheck(QuotaCheck):
     key = "elb_nlb_count"
     description = "Network Load Balancers per region"
     scope = QuotaScope.REGION
-    service_code = "elasticloadbalancing"
-    quota_code = "L-69A177A2"
+    service_code = 'elasticloadbalancing'
+    quota_code = 'L-69A177A2'
 
     @property
     def current(self):
         return len(
             list(
                 filter(
-                    lambda lb: lb["Type"] == "network",
-                    self.boto_session.client("elbv2").describe_load_balancers()["LoadBalancers"],
+                    lambda lb: lb['Type'] == 'network',
+                    self.boto_session.client('elbv2').describe_load_balancers()['LoadBalancers'],
                 )
             )
         )
@@ -84,23 +84,23 @@ class NetworkLoadBalancerCountCheck(QuotaCheck):
 class ListenerPerNetworkLoadBalancerCountCheck(InstanceQuotaCheck):
     key = "elb_listeners_per_nlb"
     description = "Listeners per Network Load Balancer"
-    service_code = "elasticloadbalancing"
-    quota_code = "L-57A373D6"
-    instance_id = "Load Balancer ARN"
+    service_code = 'elasticloadbalancing'
+    quota_code = 'L-57A373D6'
+    instance_id = 'Load Balancer ARN'
 
     @staticmethod
     def get_all_identifiers(session: boto3.Session) -> typing.List[str]:
-        return [alb["LoadBalancerArn"] for alb in get_nlbs(session)]
+        return [alb['LoadBalancerArn'] for alb in get_nlbs(session)]
 
     @property
     def current(self):
         try:
             return len(
-                self.boto_session.client("elbv2").describe_listeners(
+                self.boto_session.client('elbv2').describe_listeners(
                     LoadBalancerArn=self.instance_id
-                )["Listeners"]
+                )['Listeners']
             )
-        except self.boto_session.client("elbv2").exceptions.LoadBalancerNotFoundException as e:
+        except self.boto_session.client('elbv2').exceptions.LoadBalancerNotFoundException as e:
             raise InstanceWithIdentifierNotFound(self) from e
 
 
@@ -108,8 +108,8 @@ class ApplicationLoadBalancerCountCheck(QuotaCheck):
     key = "elb_alb_count"
     description = "Application Load Balancers per region"
     scope = QuotaScope.REGION
-    service_code = "elasticloadbalancing"
-    quota_code = "L-53DA6B97"
+    service_code = 'elasticloadbalancing'
+    quota_code = 'L-53DA6B97'
 
     @property
     def current(self):
@@ -119,23 +119,23 @@ class ApplicationLoadBalancerCountCheck(QuotaCheck):
 class ListenerPerApplicationLoadBalancerCountCheck(InstanceQuotaCheck):
     key = "elb_listeners_per_alb"
     description = "Listeners per Application Load Balancer"
-    service_code = "elasticloadbalancing"
-    quota_code = "L-B6DF7632"
-    instance_id = "Load Balancer ARN"
+    service_code = 'elasticloadbalancing'
+    quota_code = 'L-B6DF7632'
+    instance_id = 'Load Balancer ARN'
 
     @staticmethod
     def get_all_identifiers(session: boto3.Session) -> typing.List[str]:
-        return [alb["LoadBalancerArn"] for alb in get_albs(session)]
+        return [alb['LoadBalancerArn'] for alb in get_albs(session)]
 
     @property
     def current(self) -> int:
         try:
             return len(
-                self.boto_session.client("elbv2").describe_listeners(
+                self.boto_session.client('elbv2').describe_listeners(
                     LoadBalancerArn=self.instance_id
-                )["Listeners"]
+                )['Listeners']
             )
-        except self.boto_session.client("elbv2").exceptions.LoadBalancerNotFoundException as e:
+        except self.boto_session.client('elbv2').exceptions.LoadBalancerNotFoundException as e:
             raise InstanceWithIdentifierNotFound(self) from e
 
 
@@ -143,32 +143,32 @@ class TargetGroupCountCheck(QuotaCheck):
     key = "elb_target_group_count"
     description = "Target Groups per region"
     scope = QuotaScope.REGION
-    service_code = "elasticloadbalancing"
-    quota_code = "L-B22855CB"
+    service_code = 'elasticloadbalancing'
+    quota_code = 'L-B22855CB'
 
     @property
     def current(self):
-        return len(self.boto_session.client("elbv2").describe_target_groups()["TargetGroups"])
+        return len(self.boto_session.client('elbv2').describe_target_groups()['TargetGroups'])
 
 
 class TargetGroupsPerApplicationLoadBalancerCountCheck(InstanceQuotaCheck):
     key = "elb_target_groups_per_alb"
     description = "Target groups per Application Load Balancer"
-    service_code = "elasticloadbalancing"
-    quota_code = "L-822D1B1B"
-    instance_id = "Load Balancer ARN"
+    service_code = 'elasticloadbalancing'
+    quota_code = 'L-822D1B1B'
+    instance_id = 'Load Balancer ARN'
 
     @staticmethod
     def get_all_identifiers(session: boto3.Session) -> typing.List[str]:
-        return [alb["LoadBalancerArn"] for alb in get_albs(session)]
+        return [alb['LoadBalancerArn'] for alb in get_albs(session)]
 
     @property
     def current(self) -> int:
         try:
             return len(
-                self.boto_session.client("elbv2").describe_target_groups(
+                self.boto_session.client('elbv2').describe_target_groups(
                     LoadBalancerArn=self.instance_id
-                )["TargetGroups"]
+                )['TargetGroups']
             )
-        except self.boto_session.client("elbv2").exceptions.LoadBalancerNotFoundException as e:
+        except self.boto_session.client('elbv2').exceptions.LoadBalancerNotFoundException as e:
             raise InstanceWithIdentifierNotFound(self) from e